+++ docs/sigv4_post_sample.py
@@ -138,7 +138,7 @@ print('Request URL = ' + endpoint)
 print('request_parameters = ' + str(request_parameters))
 print('headers = ' + str(headers))
 
-r = requests.post(endpoint, data=request_parameters, headers=headers)
+r = requests.post(endpoint, data=request_parameters, headers=headers, timeout=60)
 
 print('\nRESPONSE++++++++++++++++++++++++++++++++++++')
 print('Response code: %d\n' % r.status_code)