+++ test/unit/anonymized-data-logger/conftest.py
@@ -5,4 +5,4 @@ import pytest
 
 @pytest.fixture(autouse=True)
 def mock_env_variables(monkeypatch):
-    monkeypatch.syspath_prepend('../../source/anonymous-data-logger/')
+    monkeypatch.syspath_prepend('../../source/anonymized-data-logger/')