+++ test/unit/operators/comprehend/entities/test_get_entity_detection.py
@@ -25,7 +25,7 @@ def stub_comprehend_client(stub):
         }
     )
 
-def stub_s3_response(stub, bucket_name, key_name, response_body):
+def stub_s3_response(stub, bucket_name: str, key_name: str, response_body: bytes):
     stub.add_response(
         'get_object',
         expected_params = {
@@ -34,11 +34,43 @@ def stub_s3_response(stub, bucket_name, key_name, response_body):
         },
         service_response = {
             'Body': StreamingBody(
-                io.BytesIO(response_body.encode()),
+                io.BytesIO(response_body),
                 len(response_body)
             )
         }
     )
+
+
+def make_tar_output(contents: str) -> bytes:
+    """Make a gzip tarball with one output file containing the provided contents.
+
+    The tarball will contain a single file called 'output' containing the specified string
+    and it will be compressed with gzip.
+
+    Args:
+        contents (str): String to compress into the output file.
+
+    Returns:
+        The bytes of the in-memory tarball.
+    """
+    # Encode the string into utf-8 bytes.
+    encoded_bytes = contents.encode()
+    # Create a buffer that tar will write to.
+    with io.BytesIO() as tarbuf:
+        # Open the buffer with tar in write mode with gzip compression.
+        with tar.open(mode='w:gz', fileobj=tarbuf) as targz:
+            # Make a buffer with the input contents to be compressed.
+            with io.BytesIO(encoded_bytes) as filebuf:
+                # Create a fake TarInfo with file name "output" and size of the contents.
+                tinfo = tar.TarInfo('output')
+                tinfo.size = len(encoded_bytes)
+                # Add the file contents to the tarball.
+                targz.addfile(tinfo, filebuf)
+                # Close the input file buffer.
+            # Close the tarball.
+        # Get the raw bytes from the output buffer before we close it. Return the bytes.
+        return tarbuf.getvalue()
+
 
 def test_parameter_validation():
     import comprehend.entities.get_entity_detection as lambda_function
@@ -158,10 +190,8 @@ def test_comprehend_job_error(comprehend_client_stub, s3_get_entity):
         s3_get_entity,
         'bucket_name',
         'comprehend/key_name',
-        'hello world'
+        make_tar_output('hello world')
     )
-    original_open = tar.open
-    tar.open = MagicMock()
 
     original_dataplane = lambda_function.DataPlane.store_asset_metadata
     lambda_function.DataPlane.store_asset_metadata = MagicMock(
@@ -172,15 +202,13 @@ def test_comprehend_job_error(comprehend_client_stub, s3_get_entity):
         response = lambda_function.lambda_handler(operator_parameter, {})
         assert response['Status'] == 'Error'
         assert 'Unable to store entity data' in response['MetaData']['comprehend_error']
-    assert tar.open.call_count == 1
     assert lambda_function.DataPlane.store_asset_metadata.call_count == 1
     assert lambda_function.DataPlane.store_asset_metadata.call_args[0][0] == 'testAssetId'
     assert lambda_function.DataPlane.store_asset_metadata.call_args[0][1] == 'entities'
     assert lambda_function.DataPlane.store_asset_metadata.call_args[0][2] == 'testWorkflowId'
-    assert lambda_function.DataPlane.store_asset_metadata.call_args[0][3] == {'LanguageCode': 'en-US', 'Results': []}
+    assert lambda_function.DataPlane.store_asset_metadata.call_args[0][3] == {'LanguageCode': 'en-US', 'Results': ['hello world']}
 
-    
-    tar.open = original_open
+
     lambda_function.DataPlane.store_asset_metadata = original_dataplane
 
 def test_comprehend_job_completed(comprehend_client_stub, s3_get_entity):
@@ -196,10 +224,8 @@ def test_comprehend_job_completed(comprehend_client_stub, s3_get_entity):
         s3_get_entity,
         'bucket_name',
         'comprehend/key_name',
-        'hello world'
+        make_tar_output('hello world')
     )
-    original_open = tar.open
-    tar.open = MagicMock()
 
     original_dataplane = lambda_function.DataPlane.store_asset_metadata
     lambda_function.DataPlane.store_asset_metadata = MagicMock(
@@ -211,13 +237,11 @@ def test_comprehend_job_completed(comprehend_client_stub, s3_get_entity):
     response = lambda_function.lambda_handler(operator_parameter, {})
     assert response['Status'] == 'Complete'
     assert response['MetaData']['output_uri'] == 's3://bucket_name/comprehend/key_name'
-    assert tar.open.call_count == 1
     assert lambda_function.DataPlane.store_asset_metadata.call_count == 1
     assert lambda_function.DataPlane.store_asset_metadata.call_args[0][0] == 'testAssetId'
     assert lambda_function.DataPlane.store_asset_metadata.call_args[0][1] == 'entities'
     assert lambda_function.DataPlane.store_asset_metadata.call_args[0][2] == 'testWorkflowId'
-    assert lambda_function.DataPlane.store_asset_metadata.call_args[0][3] == {'LanguageCode': 'en-US', 'Results': []}
+    assert lambda_function.DataPlane.store_asset_metadata.call_args[0][3] == {'LanguageCode': 'en-US', 'Results': ['hello world']}
 
-    
-    tar.open = original_open
+
     lambda_function.DataPlane.store_asset_metadata = original_dataplane