+++ CHANGELOG.md
@@ -5,6 +5,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [5.1.4] - 2023-09-21
+
+### Added
+
+* Added support for Python 3.10 and 3.11
+
+### Changed
+
+* Updated existing Lambda functions to Python 3.11 runtime and layer
+
+### Removed
+
+* Removed Python 3.7 and 3.8 references from the solution - lambda layers can still be downloaded, but they won't be used by default in the solution
+
+### Fixed
+
+* Updated Send Anonymous Data from CfnParameter to CfnMapping
+* Changed "Anonymous Metrics" to "Anonymized Metrics"
+
 ## [5.1.3] - 2023-05-01
 
 ### Changed